refactor: split TradingView settings gallery components
